Refactor infoMarcaciones function to include current date filter and order by Id
