+++ src/utils/funtions.ts
@@ -1,4 +1,6 @@
+import { Marcacion } from '../model/marcacion.model';
 import { MainError } from '../types/error.mysql';
+import { Estado } from '../types/info';
 
 /**
  * Función para validar si un Error de type unknow pertenece a Interface MainError. 
@@ -13,6 +15,27 @@ function isMainError(error: unknown): error is MainError {
     'parent' in error &&
     typeof (error as MainError).parent.code === 'string'
   );
+}
+
+export function reduceStates (marcacion: Marcacion[]) {
+
+  const estadosMap: { [key: string]: Estado } = {
+    'Entrada': Estado.Entrada,
+    'Salida_intermedia': Estado.SalidaIntermedia,
+    'Entrada_intermedia': Estado.EntradaIntermedia,
+    'Salida': Estado.Salida
+  };
+
+  return marcacion.reduce((acc, mar) => {
+    const key = estadosMap[mar.estado];
+    if (key) { acc[key] += 1; }
+    return acc
+  }, { 
+    [Estado.Entrada]: 0,
+    [Estado.SalidaIntermedia]: 0,
+    [Estado.EntradaIntermedia]: 0,
+    [Estado.Salida]: 0
+  })
 }
 
 export default isMainError;
